requirements.txt
redis==3.5.3
python-telegram-bot==13.7
notion-client==0.7.1
python-box==5.4.1
pyyaml==6.0